Fix Invalid cookie header messages
When downloading from some repositories like the oracle maven repository
httpclient is not able to parse the expires attribute in the set cookie
header.
This looks like using the `standard` cookie spec instead of the default
one fixes the problem. It is also necessary to upgrade httpclient to
the latest version to make this work for
https://issues.apache.org/jira/browse/HTTPCLIENT-1077.